Implement Role-Based Permissions, Password Features, and UI Fixes
This commit introduces several new features, bug fixes, and UX improvements to the user and task management workflows.

Features:
- **Password Management:**
  - The 'Add User' popup now includes a 'Confirm Password' field with validation.
  - The 'Edit User' popup features a 'Change Password' button to conditionally display the password fields for updates.
- **Role-Based Permissions:**
  - Implemented role-based permissions for editing tasks. Superadmins can edit all task fields, while regular users can only update the status. The backend enforces this logic.
  - A delete icon for tasks is now visible only to admin users in the tasks table.
- **Confirmation Dialogs:**
  - Added confirmation dialogs for deleting both users and tasks to prevent accidental deletions.

Fixes:
- **UI/UX:**
  - The 'Assigned To' column in the tasks table now correctly displays the username instead of the ID.
  - The 'Assigned To' dropdown in the `TaskPopup` no longer shows duplicate entries for the current user.
- **Bugs:**
  - Fixed a race condition that caused incorrect UI state for admins by adding a loading check.
  - Corrected backend logic and frontend calls to ensure the application is fully functional after the refactoring.
  - The `TaskPopup` now correctly handles its open/close state on mutation errors.
