+++ backend/apps/users/views.py
@@ -41,7 +41,7 @@ def user_list(request):
             "first_name": user.first_name,
             "last_name": user.last_name,
             "is_staff": user.is_staff,
-            "is_superadmin":user.is_superuser
+            "is_superuser":user.is_superuser
         }
         for user in users
     ]
