+++ frontend/src/types/user.ts
@@ -1,0 +1,11 @@
+export interface User {
+  id: number;
+  username: string;
+  email: string;
+  first_name: string;
+  last_name: string;
+  is_staff: boolean;
+  is_superuser: boolean;
+  groups: any[];
+  permissions: string[];
+}
